+++ hw05_parallel_execution/run_test.go
@@ -1,11 +1,9 @@
 package hw05parallelexecution
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"math/rand"
-	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -13,53 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
 )
-
-var ErrErrorsLimitExceeded = errors.New("errors limit exceeded")
-
-type Task func() error
-
-func worker(ctx context.Context, wg *sync.WaitGroup, tasks <-chan Task, errCounter *int32) {
-	defer wg.Done()
-	for {
-		select {
-		case task, ok := <-tasks:
-			if !ok {
-				return
-			}
-			if err := task(); err != nil {
-				atomic.AddInt32(errCounter, 1)
-			}
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-// Run starts tasks in n goroutines and stops its work when receiving m errors from tasks.
-func Run(tasks []Task, n, m int) error {
-	ctx, cancelFunc := context.WithCancel(context.Background())
-	tasksChain := make(chan Task)
-	defer close(tasksChain)
-	wg := &sync.WaitGroup{}
-
-	var errCounter int32 = 0
-	for i := 0; i < n; i++ {
-		wg.Add(1)
-		go worker(ctx, wg, tasksChain, &errCounter)
-	}
-
-	waitForEnd := func() { cancelFunc(); wg.Wait() }
-
-	for _, task := range tasks {
-		if int(atomic.LoadInt32(&errCounter)) >= m {
-			waitForEnd()
-			return ErrErrorsLimitExceeded
-		}
-		tasksChain <- task
-	}
-	waitForEnd()
-	return nil
-}
 
 func TestRun(t *testing.T) {
 	defer goleak.VerifyNone(t)
@@ -115,5 +66,28 @@ func TestRun(t *testing.T) {
 
 		require.Equal(t, runTasksCount, int32(tasksCount), "not all tasks were completed")
 		require.LessOrEqual(t, int64(elapsedTime), int64(sumTime/2), "tasks were run sequentially?")
+	})
+
+	t.Run("unlimited errors", func(t *testing.T) {
+		tasksCount := 50
+		tasks := make([]Task, 0, tasksCount)
+
+		var runTasksCount int32
+
+		for i := 0; i < tasksCount; i++ {
+			err := fmt.Errorf("error from task %d", i)
+			tasks = append(tasks, func() error {
+				time.Sleep(time.Millisecond * time.Duration(rand.Intn(100)))
+				atomic.AddInt32(&runTasksCount, 1)
+				return err
+			})
+		}
+
+		workersCount := 10
+		maxErrorsCount := 0
+		err := Run(tasks, workersCount, maxErrorsCount)
+
+		require.NoError(t, err)
+		require.Equal(t, runTasksCount, int32(tasksCount), "extra tasks were started")
 	})
 }
